Move `ParseInvocation` into a private namespace in `command_request.cpp`
Bug: b/382745645
